made changes for OOBA setup

+++ launch_model_ooba.sh
@@ -1,18 +1,26 @@
 #!/bin/bash
 
+deactivate
 if [ -z "$SERVER_DIR" ]
 then
     current_cwd=$(pwd)
     export SERVER_DIR="$current_cwd"
 fi
 
-pip install accelerate -U
-pip install protobuf
-python -m pip install git+https://github.com/jllllll/exllama
+if [ ! -f "/home/workspace/init_launch" ]
+then
+    pip install accelerate -U
+    pip install protobuf
+    python -m pip install git+https://github.com/jllllll/exllama
+    touch /home/workspace/init_launch
+fi
 
 MODEL_PATH="models/${MODEL_USER}_${MODEL_NAME}"
 echo "$MODEL_PATH"
 cd /src
-python3 /app/download-model.py $MODEL_USER/$MODEL_NAME > $SERVER_DIR/download.log 2>&1
+if [ ! -d "$MODEL_PATH" ]
+then
+    python3 /app/download-model.py $MODEL_USER/$MODEL_NAME > $SERVER_DIR/download.log 2>&1
+fi
 python3 /app/server.py --extensions api --api-blocking-port 5001 --api-streaming-port 5002 --model $MODEL_PATH --listen > $SERVER_DIR/infer.log 2>&1 &
 echo "done"
